--- src/main/java/nextstep/subway/applicaion/dto/SectionRequest.java
@@ -1,19 +1,0 @@
-package nextstep.subway.applicaion.dto;
-
-public class SectionRequest {
-    private Long upStationId;
-    private Long downStationId;
-    private int distance;
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-}
