--- src/main/java/nextstep/subway/applicaion/dto/StationRequest.java
@@ -1,9 +1,0 @@
-package nextstep.subway.applicaion.dto;
-
-public class StationRequest {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-}
