+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -1,0 +1,113 @@
+package nextstep.subway.line.application;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import nextstep.subway.line.domain.dto.LineRequest;
+import nextstep.subway.line.domain.dto.LineResponse;
+import nextstep.subway.line.domain.dto.SectionRequest;
+import nextstep.subway.line.domain.model.Line;
+import nextstep.subway.line.domain.model.Section;
+import nextstep.subway.line.domain.repository.LineRepository;
+import nextstep.subway.station.application.StationService;
+import nextstep.subway.station.domain.dto.StationResponse;
+import nextstep.subway.station.domain.model.Station;
+
+@Service
+@Transactional
+public class LineService {
+    private LineRepository lineRepository;
+    private StationService stationService;
+
+    public LineService(LineRepository lineRepository, StationService stationService) {
+        this.lineRepository = lineRepository;
+        this.stationService = stationService;
+    }
+
+    public LineResponse saveLine(LineRequest request) {
+        Line line = lineRepository.save(new Line(request.getName(), request.getColor()));
+        if (request.getUpStationId() != null && request.getDownStationId() != null && request.getDistance() != 0) {
+            Station upStation = stationService.findById(request.getUpStationId());
+            Station downStation = stationService.findById(request.getDownStationId());
+            line.getSections().add(new Section(line, upStation, downStation, request.getDistance()));
+        }
+        return createLineResponse(line);
+    }
+
+    @Transactional(readOnly = true)
+    public List<LineResponse> showLines() {
+        return lineRepository.findAll().stream()
+                .map(this::createLineResponse)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional(readOnly = true)
+    public LineResponse findById(Long id) {
+        return createLineResponse(lineRepository.findById(id).orElseThrow(IllegalArgumentException::new));
+    }
+
+    public void updateLine(Long id, LineRequest lineRequest) {
+        Line line = lineRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+
+        if (lineRequest.getName() != null) {
+            line.setName(lineRequest.getName());
+        }
+        if (lineRequest.getColor() != null) {
+            line.setColor(lineRequest.getColor());
+        }
+    }
+
+    public void deleteLine(Long id) {
+        lineRepository.deleteById(id);
+    }
+
+    public void addSection(Long lineId, SectionRequest sectionRequest) {
+        Station upStation = stationService.findById(sectionRequest.getUpStationId());
+        Station downStation = stationService.findById(sectionRequest.getDownStationId());
+        Line line = lineRepository.findById(lineId).orElseThrow(IllegalArgumentException::new);
+
+        line.getSections().add(new Section(line, upStation, downStation, sectionRequest.getDistance()));
+    }
+
+    private LineResponse createLineResponse(Line line) {
+        return new LineResponse(
+                line.getId(),
+                line.getName(),
+                line.getColor(),
+                createStationResponses(line),
+                line.getCreatedDate(),
+                line.getModifiedDate()
+        );
+    }
+
+    private List<StationResponse> createStationResponses(Line line) {
+        if (line.getSections().isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        List<Station> stations = line.getSections().stream()
+                .map(Section::getDownStation)
+                .collect(Collectors.toList());
+
+        stations.add(0, line.getSections().get(0).getUpStation());
+
+        return stations.stream()
+                .map(it -> stationService.createStationResponse(it))
+                .collect(Collectors.toList());
+    }
+
+    public void deleteSection(Long lineId, Long stationId) {
+        Line line = lineRepository.findById(lineId).orElseThrow(IllegalArgumentException::new);
+        Station station = stationService.findById(stationId);
+
+        if (!line.getSections().get(line.getSections().size() - 1).getDownStation().equals(station)) {
+            throw new IllegalArgumentException();
+        }
+
+        line.getSections().remove(line.getSections().size() - 1);
+    }
+}
