+++ src/main/java/nextstep/subway/line/domain/dto/LineResponse.java
@@ -1,0 +1,49 @@
+package nextstep.subway.line.domain.dto;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import nextstep.subway.station.domain.dto.StationResponse;
+
+public class LineResponse {
+    private Long id;
+    private String name;
+    private String color;
+    private List<StationResponse> stations;
+    private LocalDateTime createdDate;
+    private LocalDateTime modifiedDate;
+
+    public LineResponse(Long id, String name, String color, List<StationResponse> stations, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+        this.stations = stations;
+        this.createdDate = createdDate;
+        this.modifiedDate = modifiedDate;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
+    }
+
+    public LocalDateTime getModifiedDate() {
+        return modifiedDate;
+    }
+}
+
