+++ src/main/java/nextstep/subway/line/domain/dto/SectionRequest.java
@@ -1,0 +1,19 @@
+package nextstep.subway.line.domain.dto;
+
+public class SectionRequest {
+    private Long upStationId;
+    private Long downStationId;
+    private int distance;
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
