+++ src/main/java/nextstep/subway/station/application/StationService.java
@@ -1,0 +1,53 @@
+package nextstep.subway.station.application;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import nextstep.subway.station.domain.dto.StationRequest;
+import nextstep.subway.station.domain.dto.StationResponse;
+import nextstep.subway.station.domain.model.Station;
+import nextstep.subway.station.domain.repository.StationRepository;
+
+@Service
+@Transactional
+public class StationService {
+    private StationRepository stationRepository;
+
+    public StationService(StationRepository stationRepository) {
+        this.stationRepository = stationRepository;
+    }
+
+    public StationResponse saveStation(StationRequest stationRequest) {
+        Station station = stationRepository.save(new Station(stationRequest.getName()));
+        return createStationResponse(station);
+    }
+
+    @Transactional(readOnly = true)
+    public List<StationResponse> findAllStations() {
+        List<Station> stations = stationRepository.findAll();
+
+        return stations.stream()
+                .map(this::createStationResponse)
+                .collect(Collectors.toList());
+    }
+
+    public void deleteStationById(Long id) {
+        stationRepository.deleteById(id);
+    }
+
+    public StationResponse createStationResponse(Station station) {
+        return new StationResponse(
+                station.getId(),
+                station.getName(),
+                station.getCreatedDate(),
+                station.getModifiedDate()
+        );
+    }
+
+    public Station findById(Long id) {
+        return stationRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+    }
+}
