+++ src/main/java/nextstep/subway/station/domain/model/Station.java
@@ -1,0 +1,33 @@
+package nextstep.subway.station.domain.model;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+import nextstep.subway.common.domain.model.BaseEntity;
+
+@Entity
+public class Station extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(unique = true)
+    private String name;
+
+    public Station() {
+    }
+
+    public Station(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
