--- src/main/java/nextstep/subway/ui/PageController.java
@@ -1,13 +1,0 @@
-package nextstep.subway.ui;
-
-import org.springframework.http.MediaType;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class PageController {
-    @GetMapping(value = {"/", "/stations", "/lines", "/sections", "/path"}, produces = MediaType.TEXT_HTML_VALUE)
-    public String index() {
-        return "index";
-    }
-}
