+++ src/test/java/nextstep/subway/acceptance/line/LineAcceptanceTest.java
@@ -1,0 +1,138 @@
+package nextstep.subway.acceptance.line;
+
+import static nextstep.subway.acceptance.line.LineSteps.*;
+import static org.assertj.core.api.Assertions.*;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.acceptance.AcceptanceTest;
+
+@DisplayName("지하철 노선 관리 기능")
+class LineAcceptanceTest extends AcceptanceTest {
+    /**
+     * When 지하철 노선 생성을 요청 하면
+     * Then 지하철 노선 생성이 성공한다.
+     */
+    @DisplayName("지하철 노선 생성")
+    @Test
+    void createLine() {
+        // when
+        ExtractableResponse<Response> response = 지하철_노선_생성_요청("2호선", "green");
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    /**
+     * Given 지하철 노선 생성을 요청 하고
+     * Given 새로운 지하철 노선 생성을 요청 하고
+     * When 지하철 노선 목록 조회를 요청 하면
+     * Then 두 노선이 포함된 지하철 노선 목록을 응답받는다
+     */
+    @DisplayName("지하철 노선 목록 조회")
+    @Test
+    void getLines() {
+        // given
+        지하철_노선_생성_요청("2호선", "green");
+        지하철_노선_생성_요청("3호선", "orange");
+
+        // when
+        ExtractableResponse<Response> response = 지하철_노선_목록_조회_요청();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+        assertThat(response.jsonPath().getList("name")).contains("2호선", "3호선");
+    }
+
+    /**
+     * Given 지하철 노선 생성을 요청 하고
+     * When 생성한 지하철 노선 조회를 요청 하면
+     * Then 생성한 지하철 노선을 응답받는다
+     */
+    @DisplayName("지하철 노선 조회")
+    @Test
+    void getLine() {
+        // given
+        ExtractableResponse<Response> createResponse = 지하철_노선_생성_요청("2호선", "green");
+
+        // when
+        ExtractableResponse<Response> response = 지하철_노선_조회_요청(createResponse);
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+        assertThat(response.jsonPath().getString("name")).isEqualTo("2호선");
+    }
+
+    /**
+     * Given 지하철 노선 생성을 요청 하고
+     * When 지하철 노선의 정보 수정을 요청 하면
+     * Then 지하철 노선의 정보 수정은 성공한다.
+     */
+    @DisplayName("지하철 노선 수정")
+    @Test
+    void updateLine() {
+        // given
+        ExtractableResponse<Response> createResponse = 지하철_노선_생성_요청("2호선", "green");
+
+        // when
+        Map<String, String> params = new HashMap<>();
+        params.put("color", "red");
+        ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .body(params)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().put(createResponse.header("location"))
+                .then().log().all().extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    /**
+     * Given 지하철 노선 생성을 요청 하고
+     * When 생성한 지하철 노선 삭제를 요청 하면
+     * Then 생성한 지하철 노선 삭제가 성공한다.
+     */
+    @DisplayName("지하철 노선 삭제")
+    @Test
+    void deleteLine() {
+        // given
+        ExtractableResponse<Response> createResponse = 지하철_노선_생성_요청("2호선", "green");
+
+        // when
+        ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .when().delete(createResponse.header("location"))
+                .then().log().all().extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    /**
+     * Given 지하철 노선 생성을 요청 하고
+     * When 같은 이름으로 지하철 노선 생성을 요청 하면
+     * Then 지하철 노선 생성이 실패한다.
+     */
+    @DisplayName("중복이름으로 지하철 노선 생성")
+    @Test
+    void duplicateName() {
+        // given
+        지하철_노선_생성_요청("2호선", "green");
+
+        // when
+        ExtractableResponse<Response> createResponse = 지하철_노선_생성_요청("2호선", "green");
+
+        // then
+        assertThat(createResponse.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+    }
+}
