+++ src/test/java/nextstep/subway/acceptance/station/StationAcceptanceTest.java
@@ -1,0 +1,99 @@
+package nextstep.subway.acceptance.station;
+
+import static nextstep.subway.acceptance.station.StationSteps.*;
+import static org.assertj.core.api.Assertions.*;
+
+import java.util.List;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.acceptance.AcceptanceTest;
+
+@DisplayName("지하철역 관리 기능")
+class StationAcceptanceTest extends AcceptanceTest {
+    /**
+     * When 지하철역 생성을 요청 하면
+     * Then 지하철역 생성이 성공한다.
+     */
+    @DisplayName("지하철역 생성")
+    @Test
+    void createStation() {
+        // when
+        ExtractableResponse<Response> response = 지하철역_생성_요청("강남역");
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        assertThat(response.header("Location")).isNotBlank();
+    }
+
+    /**
+     * Given 지하철역 생성을 요청 하고
+     * Given 새로운 지하철역 생성을 요청 하고
+     * When 지하철역 목록 조회를 요청 하면
+     * Then 두 지하철역이 포함된 지하철역 목록을 응답받는다
+     */
+    @DisplayName("지하철역 목록 조회")
+    @Test
+    void getStations() {
+        // given
+        지하철역_생성_요청("강남역");
+        지하철역_생성_요청("역삼역");
+
+        // when
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .when()
+                .get("/stations")
+                .then().log().all()
+                .extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+        List<String> stationNames = response.jsonPath().getList("name");
+        assertThat(stationNames).contains("강남역", "역삼역");
+    }
+
+    /**
+     * Given 지하철역 생성을 요청 하고
+     * When 생성한 지하철역 삭제를 요청 하면
+     * Then 생성한 지하철역 삭제가 성공한다.
+     */
+    @DisplayName("지하철역 삭제")
+    @Test
+    void deleteStation() {
+        // given
+        ExtractableResponse<Response> createResponse = 지하철역_생성_요청("강남역");
+
+        // when
+        String uri = createResponse.header("Location");
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+                .when()
+                .delete(uri)
+                .then().log().all()
+                .extract();
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+
+    /**
+     * Given 지하철역 생성을 요청 하고
+     * When 같은 이름으로 지하철역 생성을 요청 하면
+     * Then 지하철역 생성이 실패한다.
+     */
+    @DisplayName("중복이름으로 지하철역 생성")
+    @Test
+    void duplicateName() {
+        // given
+        지하철역_생성_요청("강남역");
+
+        // when
+        ExtractableResponse<Response> createResponse = 지하철역_생성_요청("강남역");
+
+        // then
+        assertThat(createResponse.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+    }
+}
