+++ src/test/java/nextstep/subway/unit/line/application/LineServiceMockTest.java
@@ -1,0 +1,29 @@
+package nextstep.subway.unit.line.application;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import nextstep.subway.line.domain.repository.LineRepository;
+import nextstep.subway.station.application.StationService;
+
+@ExtendWith(MockitoExtension.class)
+public class LineServiceMockTest {
+    @Mock
+    private LineRepository lineRepository;
+    @Mock
+    private StationService stationService;
+
+    @Test
+    void addSection() {
+        // given
+        // lineRepository, stationService stub 설정을 통해 초기값 셋팅
+
+        // when
+        // lineService.addSection 호출
+
+        // then
+        // line.findLineById 메서드를 통해 검증
+    }
+}
