+++ src/test/java/nextstep/subway/unit/line/domain/LineTest.java
@@ -1,0 +1,21 @@
+package nextstep.subway.unit.line.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class LineTest {
+    @DisplayName("구간 목록 마지막에 새로운 구간을 추가할 경우")
+    @Test
+    void addSection() {
+    }
+
+    @DisplayName("노선에 속해있는 역 목록 조회")
+    @Test
+    void getStations() {
+    }
+
+    @DisplayName("구간이 목록에서 마지막 역 삭제")
+    @Test
+    void removeSection() {
+    }
+}
